M13-PROJECT/PycharmProjects/dam2-m13/src/python/08-hangman/requirements.txt
pyinstaller~=6.0.0
setuptools~=68.2.2
requests~=2.31.0
psycopg2-binary~=2.9.9